Update server.py
When you type the command "say Hello! Is there anyone here?" and hit enter, the client sends that enter key on the end of your command. And then it gets put on the end of your words that your character is saying.

This is a good place to go ahead and strip that extra newline off. :)